Initial changes for new Question
