+++ components/QuestionEditCard/QuestionEditCard.module.scss
@@ -1,4 +1,4 @@
-.questionEdit {
+.questionEditCard {
   display: flex;
   justify-content: space-between;
   align-items: center;
