+++ messages/en-US/global.json
@@ -125,7 +125,8 @@
     "messaging": {
       "loading": "Loading",
       "noItemsFound": "No items found",
-      "somethingWentWrong": "Something went wrong. Please try again."
+      "somethingWentWrong": "Something went wrong. Please try again.",
+      "resultsText": "{name} results match your search"
     }
   }
 }
